+++ pkg/reconcilers/marin3r/envoybootstrap/client_certificate.go
